feat: Modularize-collections-#24 - closes #24
